+++ packages/common/src/permissions/types/Permission.ts
@@ -32,7 +32,9 @@ const SystemPermissions = [
   // remove this at the next breaking version
   "hub:feature:workspace:umbrella",
   "hub:feature:keyboardshortcuts",
+  "hub:feature:newentityview",
   "hub:feature:history",
+  "hub:feature:catalogs",
   "hub:license:hub-premium",
   "hub:license:hub-basic",
   "hub:license:enterprise-sites",