+++ packages/common/src/projects/_internal/ProjectBusinessRules.ts
@@ -36,6 +36,7 @@ export const ProjectPermissions = [
   "hub:project:workspace:content",
   "hub:project:workspace:events",
   "hub:project:workspace:metrics",
+  "hub:project:workspace:catalogs",
   "hub:project:manage",
 ] as const;
 
@@ -164,6 +165,14 @@ export const ProjectPermissionPolicies: IPermissionPolicy[] = [
   {
     permission: "hub:project:workspace:metrics",
     dependencies: ["hub:project:workspace", "hub:project:edit"],
+  },
+  {
+    permission: "hub:project:workspace:catalogs",
+    dependencies: [
+      "hub:project:workspace",
+      "hub:feature:catalogs",
+      "hub:project:edit",
+    ],
   },
   {
     permission: "hub:project:manage",